// src/tabs/Tab.d.ts
import { TabGroupProps, TabListProps, TabPanelProps, TabPanelsProps, TabProps } from './tab.interface';

declare const Tab: {
    ({ disabled, children }: TabProps): import("react/jsx-dev-runtime").JSX.Element;
    Group: ({ children }: TabGroupProps) => import("react/jsx-dev-runtime").JSX.Element;
    List: ({ children }: TabListProps) => import("react/jsx-dev-runtime").JSX.Element;
    Panel: ({ children }: TabPanelProps) => import("react/jsx-dev-runtime").JSX.Element;
    Panels: ({ children }: TabPanelsProps) => import("react/jsx-dev-runtime").JSX.Element;
};
export { Tab };
//# sourceMappingURL=Tab.d.ts.map